Basic terrain with AJ character from Mixamo
Basic terrain with part road layout, castle bridge, and river channel. and a few warehouse buildings.

Updated gitignore file to prevent inclusion of "TerrainSampleAssets" and any other used assets from the Asset Store, BUT to include ALL meta files (whether or not they are needed) - I think this is correct!